Reduce modularisation in AdvancedOptions
