--- lib/ReactViews/Map/Panels/SharePanel/AdvancedOptions/Embed/index.ts
@@ -1,1 +1,0 @@
-export { EmbedSection } from "./EmbedSection";
